Deploy a quickstart configmap alongside tap-telemetry
After the initial TAP data values have been created, and just before
we install TAP, we can do 2 things:
- we can deploy some secrets into the tap-install namespace to later act
  as ytt overlays for the packages
- we can patch the TAP data values, to add those secrets as package
  overlays

With this approach we don't have to care about much. Any additionaly
created resource will be managed, as part of the package install, by the
kapp-controller. It will be cleaned up when the packages is deleted, it
will be reconciled in case it was changed/deleted.

The first candidate for this is a new config map we need to deploy
alongside tap-telemetry.

For any other addition like this one, we need to:
- ensure the overlay is deployed as a secret into the tap-install
  namespace; thus we need to add one or more secrets into
  .../resources/package_overlay_secrets.yml
- we need to ensure this overlay get configured in TAP's data values,
  thus add new things to the patch-overlay in
  .../resources/package_overlay_apply.yml