Cleanup qs-test's vars
We don't need the `tappc-registry` in the the `qs-test` pipeline right
now, we have that in the task files.
Once we need more control over the actual version of the image we use,
we can either use the task's `image` stanza or use `vars` to get the
actual image version configured.
For now, let's roll with hardcoded, thus `latest`. #yolo