Add some static checks
For now, only one check on our template file is implemented. This seems
easy enough to run without a bigger testing framework. However, we might
change our opinion on that later.

This test does not run in the pipeline yet. I'll spin out a story to get
this test (and others) implemented.