Add PAT expiration check
Will run as part of the housekeeping job and notify us via slak if it
fails.

I didn't find an API where we can check the expiration time of a token.
So the next best thing I came up with is to update the expiration date
in the secret itself, alongside the token, and check on that.