Signal result of tap-install-script back to the stack
A couple of things are happening here:
1. We define some WaitConditions, for now
   - PhaseCloudInit
   - PhaseTAPInstall
   - PhaseTAPWorkloadInstall
   - PhaseTAPTests
   which get created at different times in the stack. E.g. the
   `PhaseCloudInit` gets created once the EC2 instance changed to `CREATE_COMPLETE`,
   however the `PhaseTAPWorkloadInstall` gets created once the
   `PhaseTAPInstall` goes into `CREATE_COMPLETE` (and thus signals that
   TAP installation has finished).
2. In the cloud-init we use `cfn-signal` to signal back success or
   failures to those WaitConditions respectively their related Handles,
   thus changing the status of those resources to `CREATE_COMPLETE` or
   `CREATE_FAILED`.
3. With that we can measure how long inidividual parts took to bring up.
   There is a task in the pipeline which pulls out some data about that.
   Once we agree on which data we should pull out, we can then also push
   that "somewhere" instead of only printing it.

Now to make this signaling via `cfn-signal` work, we need to pass in the
`WaitConditionHandle` URLs into the instance's user-data. With those
URLs being quite big, the maximum size of the user-data was exceeded.
Thus we pulled out quite some stuff into a separate script, which we
- pull down from AWS
- launch from the inline user-data block

For now, the general approach is:
Inline pretty much only things we need to pull down the externalized
cloud-init script (package and CLI installation), the setup of the
signaling handler/trap, and exporting of variables the externalized
cloud-init script needs.
The rest, pretty much everything related to setup and configuration of
TAP itself, should go to the externalized cloud-init script.

This also has the side-effect, that we don't have to deal with
shell-vars vs. cloudformation-subtitution-vars in the externalized
script; we only have to deal with that in the inlined user-data.