Disable comments on PR status updates
Right now the comments look like as if they came from the person who
created the PAT for the concourse pr-resoures (right now: me). That's a
bit weird. To change that, we can look into:
- creating a GitHub account (e.g. `tappc-robot` or so) and use a PAT
  from that account
- Creating a GitHub application and make the concourse pr-resource use
  that

For now, I think it's good enough to just set the context but don't
comment on the PR.