Change payload for cfn-signal
Turns out, the data we send with `--data` is nowhere to be found.

However, data we send via `--reason` is added to the stack event's
status reason.

So sending `--reason "some reason"` and a failure results in a status
reason like:
```
WaitCondition received failed message: 'some reason' for uniqueId: i-06338885b90aa37d9
```

Thus we'll use this to send the last couple of cloud init logs to give
users some context on what went wrong.

Note:
We won't see the newlines in the WebUI, however we see them in the API:
```
: aws cf describe-stack-events --region us-east-1 --stack-name "${VPCStackName}" | jq '.StackEvents[0].ResourceStatusReason' -r
WaitCondition received failed message: '
>>> package(namespace-provisioner) failed to reconcile (Reconciling), waiting for reconcile
>>> package(scanning) failed to reconcile (Reconciling), waiting for reconcile
>>> package(service-bindings) failed to reconcile (Reconciling), waiting for reconcile
>>> package(tap) failed to reconcile (Reconciling), waiting for reconcile
>>> package(tap) failed to reconcile (Reconciling), waiting for reconcile
>>> package(tap-auth) failed to reconcile (Reconciling), waiting for reconcile
>>> package(tap-telemetry) failed to reconcile (Reconciling), waiting for reconcile
>>> package(tekton-pipelines) failed to reconcile (Reconciling), waiting for reconcile
Terminated
' for uniqueId: i-0a7ca048976149f86
:
```