Run the tests against `main` daily
We have to pass through the daily trigger to the test run jobs for
concourse to actually run them daily, otherwise it would only prepare
the bucket and stop there.

We don't pass the daily trigger to the publish-bucket job, tho; we
really only need to do that when the repo's version changed.

So now, the test really run against repo's `main`, so you could consider
that a soak test.