Introduce some tests in the pipeline before we kick off e2e
This, for now, just runs some linting and other static checks.

In the future we will probably introduce more tests (e.g.
inclusive-lang, shellcheck, unit tests, ...) which should probably run
as additional `test-...` jobs in parallel to the `test-static` job, so
that they can all run and be re-triggered in isolation.