Adapt to changed test names
Now that we use the "main" taskcat config, instead of a separate one for
CI, we need to adapt to the slightly different test names.

Side-note: We could also dynamically pull out the test names from the
taskcat config, run load_vars, and the across step ... but that's for
another day.