Ensure we pass all needed values to the TAP setup script
On a refactoring of the user-data/cloud-init things some configs have
been removed and not passed on to the TAP setup script anymore by error.

This should fix this for those two variables
- `AWSAccountID`
- `TAPWorkloadBundleRepo_RepositoryUri`

In addition, to make this fail in future we now bail out from the TAP
install script if variables are referenced which are not set.