Adapt timeouts for relocation
Relocating images from the VMware repos to the internal ECR repos takes
time, we need to account for that.

To make this somewhat dynamic, based on both the cluster architecture
and relocation, we need to go through two maps. We sadly can't do
something like
```
  Timeout: !FindInMap [Timeouts, !Sub '${TAPClusterArch}-${TanzuNetRelocateImages}', CloudInit]
```
because in `Fn::FindInMap` we can only use the intrinsic functions
`Fn::Ref` and `Fn::FindInMap`, but not any others.